removed unused uploadRouter, moved current uploadToS3 to middleware; moved OLD uploadToS3 to Lab18_unusedcode
