+++ Lab18_unusedcode/uploadToS3Labs18.js
@@ -2,20 +2,24 @@ const AWS = require("aws-sdk");
 const express = require("express");
 
 const chapterDB = require("../models/chapters-model.js");
-
+//TODO Need keys for AWS up-loader
 // CONSTANTS
 const S3_BUCKET_NAME = process.env.S3_BUCKET_NAME;
 const S3_SECRET_ACCESS_KEY = process.env.S3_SECRET_ACCESS_KEY;
 const S3_ACCESS_KEY_ID = process.env.S3_ACCESS_KEY_ID;
 
+// UPLOAD TO S3 BUCKET
+
 const uploadToS3 = (file, res) => {
+  //this enables access to the S3 bucket for the project. Keys can be found inside the environment variables in Heroku
   const s3Bucket = new AWS.S3({
     accessKeyId: S3_ACCESS_KEY_ID,
     secretAccessKey: S3_SECRET_ACCESS_KEY,
     Bucket: S3_BUCKET_NAME
   });
 
-  const uploadData = s3Bucket.createBucket(() => {
+  //this creates the bucket/space within the s3bucket for the upload
+  s3Bucket.createBucket(() => {
     const params = {
       Bucket: S3_BUCKET_NAME,
       Key: file.name,
@@ -23,15 +27,18 @@ const uploadToS3 = (file, res) => {
       Body: file.data
     };
 
+    // call S3 to upload file to bucket
+
     s3Bucket.upload(params, (err, data) => {
       if (err) {
-        console.log("error in callback");
-        console.log(err);
+        // console.log("error in callback");
+        // console.log(err);
         return err;
       }
-      console.log("success");
+      // console.log("success");
 
       return data;
+
       // res.status(200).json(data);
     });
   });
